requirements.txt
# 환경구성
Django==5.0.3
djangorestframework==3.14
asgiref==3.7.2
pytz==2023.3
sqlparse==0.4.4
gunicorn==21.2.0

# debug
django-debug-toolbar==4.3.0

# db
psycopg2==2.9.6

# 인증
# djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0

# test
pytest==7.4.0
pytest-django==4.5.2

# formatter
black==23.7.0
flake8==6.1.0
